moved Hunter into own files
